refactor: reorganises monolith codebase
