+++ packages/slackBotFunction/app/core/config.py
@@ -1,0 +1,48 @@
+"""
+Core configuration for the Slack bot.
+Sets up all the AWS and Slack connections we need.
+"""
+
+import os
+import json
+import boto3
+from slack_bolt import App
+from aws_lambda_powertools import Logger
+from aws_lambda_powertools.utilities.parameters import get_parameter
+
+
+# set up logging
+logger = Logger(service="slackBotFunction")
+
+# DynamoDB table for deduplication and session storage
+dynamodb = boto3.resource("dynamodb")
+table = dynamodb.Table(os.environ["SLACK_BOT_STATE_TABLE"])
+
+# get Slack credentials from Parameter Store
+bot_token_parameter = os.environ["SLACK_BOT_TOKEN_PARAMETER"]
+signing_secret_parameter = os.environ["SLACK_SIGNING_SECRET_PARAMETER"]
+
+bot_token_raw = get_parameter(bot_token_parameter, decrypt=True)
+signing_secret_raw = get_parameter(signing_secret_parameter, decrypt=True)
+
+# parse the JSON stored in parameter store
+bot_token = json.loads(bot_token_raw)["token"]
+signing_secret = json.loads(signing_secret_raw)["secret"]
+
+# initialise the Slack app
+app = App(
+    process_before_response=True,
+    token=bot_token,
+    signing_secret=signing_secret,
+)
+
+# Bedrock configuration from environment
+KNOWLEDGEBASE_ID = os.environ["KNOWLEDGEBASE_ID"]
+RAG_MODEL_ID = os.environ["RAG_MODEL_ID"]
+AWS_REGION = os.environ["AWS_REGION"]
+GUARD_RAIL_ID = os.environ["GUARD_RAIL_ID"]
+GUARD_VERSION = os.environ["GUARD_RAIL_VERSION"]
+
+logger.info(f"Guardrail ID: {GUARD_RAIL_ID}, Version: {GUARD_VERSION}")
+
+# event handlers will be registered in main.py to avoid circular imports
